feat: add Jupiter integrated terminal in swap view
